Characterisation of allocated memory in libxmem
Add a new parameter to xmalloc(), which is an arbitrary text that
becomes attached to the memory block. The accounting functions then
can use it to investigate the source of these.

The main idea is to be able to quickly tell where lost memory comes
from.

The character() function is added in order to be able to get
information about the blocks that are handled.