Fix a bug introduced in xstrndup() with zeroing
When memory zeroing on xfree() was enabled, a piece of memory got
with xstrndup() could zero out more than actually stored, since
the size stored was actually the max size instead of the length of
the string. This happened when the length of the string was lesser
than the max len passed (practically in all cases).