Memory log as function, disable zeroing
As libxmem is a separate function, it's desirable to provide a
function for memory log that can be toggled live, since at the
time of libxmem build we won't know whether the client will want
it or not.

Enable a new function xmem_enable_memlog() that enables the
functionality in libxmem.

Disable zeroing, since it's a dubious practice in modern OSes.
Reading from the memory left behind is a segmentation fault in
many archs, so it's difficult to test or can't be used reliably.